refactor(lnswap): take the swap client from @arkade-os/swap, drop the copy
`src/lib/arkadeSwap/` held rfq.ts, onchainHtlc.ts and their program
artifact, copied out of the ts-sdk monorepo because the package they
belong to was not installable: it declared `@arkade-os/sdk` as
`workspace:*`, which only resolves inside that monorepo. ts-sdk
feat/arkade-swap names a published SDK version instead, so pnpm can now
take the package straight from the branch and the copy has no reason to
exist.

The seam lnSwap.ts was built around held: repointing the imports is the
whole change, and the code behind them is byte-identical to what was
vendored, so no behaviour moves with it.

Two things worth noting for review:

- `arklabs-lightning.card.json` is NOT part of the package — it is this
  wallet's bundled solver card, and it only ever sat in that directory
  by proximity. It moves to src/lib/swap/ alongside the code that reads
  it, and scripts/probe-rfq-solver.mjs follows it.
- .prettierignore and .eslintignore existed solely to hold the copy to
  the SDK's formatting. Both files are now empty of purpose, so both go;
  the package's own repo lints it from here on.

The dependency is a git reference until @arkade-os/swap reaches npm, so
pnpm builds it from source on install (its prepack runs tsup). A git
dependency may only run build scripts if it is allowlisted, hence the
allowBuilds entry in pnpm-workspace.yaml.
