+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@arkade-os/solver-discovery": "0.2.1",
     "@arkade-os/sdk": "0.4.56",
+    "@arkade-os/swap": "github:arkade-os/ts-sdk#feat/arkade-swap&path:/packages/swap",
     "@base-ui/react": "^1.4.1",
     "@branta-ops/branta": "3.1.3",
     "@fontsource-variable/geist": "^5.2.8",
