+++ src/test/lib/lnSwap.test.ts
@@ -1,3 +1,4 @@
+import { RFQ_TERMINAL_STATES } from '@arkade-os/swap'
 import { describe, it, expect } from 'vitest'
 import {
   InvoiceRejected,
@@ -7,7 +8,6 @@ import {
   toInvoiceFacts,
   type SwapStatusUI,
 } from '../../lib/lnSwap'
-import { RFQ_TERMINAL_STATES } from '../../lib/arkadeSwap/rfq'
 import fixtures from '../fixtures.json'
 
 describe('lnSwap', () => {
